separated testcontainer created first repo test merged with dev copy and fix issues
